Move VoyagerData class to _core/voyagerpy.py
